bot-t800-poc/requirements.txt
requests==2.23.0
botbuilder-integration-aiohttp>=4.11.0
keyring>=23.0.1
gTTS>=2.2.2
pydub>=0.25.1
python-jose[cryptography]==3.2.0
